Physics: Vectors are now arrays; vector and scalar are generic over T
I also added a `Field` trait which may eventually become its own struct.
The idea is that it keeps us from having to write out all the
contraintes everytime.

Liam seems to want to make the scalars and Vectors traits that can be
applied to any collection. This requires some thought.